+++ frontend/src/components/messstelle/MessstelleOverviewPanel.vue
